+++ app/src/main/java/eu/kanade/presentation/library/components/LibraryDisplaySettings.kt
@@ -1,6 +1,7 @@
 package eu.kanade.presentation.library.components
 
 import androidx.compose.runtime.Immutable
+import tachiyomi.domain.library.model.LibraryPinnedDisplayStyle
 
 @Immutable
 data class LibraryDisplaySettings(
@@ -9,4 +10,5 @@ data class LibraryDisplaySettings(
     val localBadge: Boolean = true,
     val languageBadge: Boolean = false,
     val entryTypeBadge: Boolean = true,
+    val pinnedDisplayStyle: LibraryPinnedDisplayStyle = LibraryPinnedDisplayStyle.TonalGroup,
 )